tidying up a bit
